move invokeClassifier to InferenceService(s)
