+++ models/mistralService.go
@@ -17,7 +17,7 @@ type mistral struct {
 	host  manifest.Host
 }
 
-func (llm *mistral) InvokeClassifier(ctx context.Context, input []string) (map[string]float64, error) {
+func (llm *mistral) InvokeClassifier(ctx context.Context, input map[string]string) (map[string]map[string]float32, error) {
 	return nil, fmt.Errorf("invokeClassifier not implemented for mistral model")
 }
 func (llm *mistral) ComputeEmbedding(ctx context.Context, sentenceMap map[string]string) (map[string][]float64, error) {
